requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs>=18.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
datasets==2.20.0
dill==0.3.8
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.23.4
idna==3.7
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.2.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
PyLaTeX==1.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xxhash==3.4.1
yarl==1.9.4
docling>=2.8
